添加电量检测功能

+++ app/src/main/java/pro/lovexj/location/util/RestartAPP.java
@@ -13,18 +13,16 @@ public class RestartAPP {
      * @param Delayed 延迟多少毫秒
      */
     public static void restartAPP(Context context, long Delayed){
-        /**
-         * 发送邮件通知
-         */
-
         /**开启一个新的服务，用来重启本APP*/
-        Intent intent1=new Intent(context, KillSelfService.class);
-        intent1.putExtra("PackageName",context.getPackageName());
-        intent1.putExtra("Delayed",Delayed);
-        context.startService(intent1);
+        if(Constant.isRestartApp){
+            Intent intent1=new Intent(context, KillSelfService.class);
+            intent1.putExtra("PackageName",context.getPackageName());
+            intent1.putExtra("Delayed",Delayed);
+            context.startService(intent1);
 
-        /**杀死整个进程**/
-        android.os.Process.killProcess(android.os.Process.myPid());
+            /**杀死整个进程**/
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
     }
     /***重启整个APP*/
     public static void restartAPP(Context context){
